SELinux requires a :Z on volumes to correctly map id/permissions between containers and host
Updating entrypoint to no longer set smtp.mock (which doesn't look like
it is ever used/referenced anywhere) and only updates the conf file if
initialAdmins or smtp.host are set (and therefore not overwriting any
values that may have been set in custom.conf already). When those
variables are set, it should work as before.