Track name as uploaded.
Changes to keep the original filename as a separate item from the
current filename for files. Added to search as a 'created_as' string
field. Default is "" and it is populated upon any call to rename(). The
display shows the new field only if it has a non zero length value.